Move hash specialization
